FIX(client): Fix rare double delete bug
AudioOutput keeps track of users and their speech in qmOutputs
The mix() method will iterate over the outputs and append
all completed outputs to a list 'qlDel'. This list is then - outside
of any synchronization - used to remove entries from qmOutputs.
Previously, other methods were also deleting from qmOutputs directly.
In rare conditions, when the output is just finished and added to qlDel
and then also manually deleted, the pointer could be subject to a double
delete.

This commit tries to mitigate the problem and overhauls the output
deletion by delegating the deletion to slots in the audio thread.