FIX(a11y): Fix all buddy labels
Qt offers a feature called buddy labels. With that certain
UI elements can have a label, which represents the UI element
in a textual way. This is very important for screen readers
as they otherwise might not read the correct for any given UI element.
Adding buddies is as simple as using the "Edit buddy" dialog
in the Qt editor.

This commit sweeps through (almost) all UI elements and
applies the correct buddies where applicable.