FIX(a11y): Improve Accessibility for CertWizard
The certificate wizard shows boxes with certificate info.
These boxes are created by using labels in a QGroupBox.
However, this information is not easily readable using a
screen reader (especially Orca).

This commit introduces the AccessibleGroupBox class
which aims to solve that by dynamically setting the
accessible description using layout information of
the labels inside the box.

Fixes #1337