FIX(a11y): Make AccessibleQGroupBox not read empty table cells
An AccessibleQGroupBox will try to compose an accessible description
based on the layout of labels within.
If a grid-layout is used, we read every row+column combination and
the corresponding grid cell content.

Previously, the AccessibleQGroup box would read row+column even if
there was not label in the actual cell, leading to unnecessary
text. (e.g. when using a sub headline)

This commit adds a check to skip a row+colum combination when
there is no label in the cell.